* Save now instantly applies without a restart.  * Workaround for "not yet loadable" albums
I figured out a way to get way more results. The auth-token is bound to a country and it's catalog. So when i manage to add accounts from different regions and merge the results, i could theoretically grab data from all catalogs. :D
I will implement this in my next commit or Qabuze 2.

Signed-off-by: Hendrik 'Xendo' Meyer <butterkekstorte@users.noreply.github.com>